MIREX2018_v1/requirements.txt
numpy==1.14.0
tensorflow-gpu==1.4.0
